requirements.txt
acme==1.1.0
arabic-reshaper==3.0.0
arrow==1.3.0
asgiref==3.8.1
asn1crypto==1.5.1
attrs==24.2.0
autobahn==17.10.1
Automat==0.8.0
backports.zoneinfo==0.2.1
bcrypt==4.2.0
blinker==1.8.2
Brotli==1.1.0
cbor==1.0.0
certbot==0.40.0
certbot-nginx==0.40.0
certifi==2024.7.4
cffi==1.17.1
channels==4.0.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
cloud-init==23.3.3
colorama==0.4.3
command-not-found==0.3
ConfigArgParse==1.7
configobj==5.0.6
constantly==15.1.0
contourpy==1.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==43.0.1
cssselect2==0.7.0
cycler==0.12.1
Cython==0.29.14
daphne==2.4.1
dbus-python==1.2.16
distlib==0.3.8
distro==1.4.0
distro-info==0.23+ubuntu1.1
Django==4.2.9
django-cors-headers==4.4.0
django-environ==0.11.2
django-extensions==3.2.3
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
django-seed==0.3.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
drf-yasg==1.21.7
entrypoints==0.3
et-xmlfile==1.1.0
Faker==26.3.0
filelock==3.13.1
fire==0.5.0
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Login==0.6.3
fonttools==4.43.1
fsspec==2023.12.2
future==0.18.2
gevent==24.2.1
geventhttpclient==2.0.12
greenlet==3.1.1
gunicorn==23.0.0
h11==0.9.0
html5lib==1.1
httplib2==0.14.0
httptools==0.1.2
hyperlink==19.0.0
idna==3.7
importlib-resources==6.1.1
importlib_metadata==8.5.0
incremental==16.10.1
inflection==0.5.1
isodate==0.6.1
itsdangerous==2.2.0
itypes==1.2.0
Jinja2==3.1.4
josepy==1.2.0
jsonpatch==1.22
jsonpointer==2.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keyring==18.0.1
kiwisolver==1.4.5
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
locust==2.25.0
lxml==5.2.2
lz4==3.0.2+dfsg
MarkupSafe==2.1.5
matplotlib==3.7.4
mock==3.0.5
more-itertools==4.2.0
mpmath==1.3.0
msgpack==1.1.0
mssql-django==1.5
netifaces==0.10.4
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.1.0
openapi-codec==1.3.2
opencv-python==4.8.1.78
openpyxl==3.1.5
oscrypto==1.3.0
packaging==24.1
paho-mqtt==2.1.0
pandas==2.0.3
parsedatetime==2.4
pbr==5.4.5
pdf2docx==0.5.6
pexpect==4.6.0
pillow==10.2.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.2
psutil==5.9.7
psycopg==3.2.3
psycopg2==2.9.9
py-cpuinfo==9.0.0
py-ubjson==0.14.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.22
PyGObject==3.36.0
PyHamcrest==1.9.0
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
PyICU==2.4.2
PyJWT==2.9.0
pymacaroons==0.13.0
PyMuPDF==1.23.5
PyMuPDFb==1.23.5
PyNaCl==1.3.0
pyodbc==5.1.0
pyOpenSSL==19.0.0
pyparsing==3.1.1
pypdf==5.0.1
pypng==0.0.20
PyQRCode==1.2.1
pyRFC3339==1.1
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.1+ubuntu0.20.4.1
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-debian==0.1.36+ubuntu1.1
python-docx==1.0.1
python-snappy==0.5.3
PyTrie==0.2
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.2.0
qrcode==7.4.2
referencing==0.35.1
reportlab==4.2.5
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
requests-unixsocket==0.2.0
roundrobin==0.0.4
rpds-py==0.20.0
scipy==1.10.1
seaborn==0.13.1
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.19.3
six==1.16.0
sos==4.5.6
sqlparse==0.5.1
ssh-import-id==5.10
supervisor==4.1.0
svglib==1.5.1
sympy==1.12
systemd-python==234
termcolor==2.3.0
testresources==2.0.1
thop==0.1.1.post2209072238
tinycss2==1.3.0
tomli==2.0.2
toposort==1.10
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
triton==2.1.0
Twisted==18.9.0
txaio==2.10.0
types-python-dateutil==2.9.0.20241003
typing_extensions==4.8.0
tzdata==2024.1
tzlocal==5.2
u-msgpack-python==2.1
ubuntu-advantage-tools==8001
ufw==0.36
ultralytics==8.0.235
unattended-upgrades==0.1
uritemplate==4.1.1
uritools==4.0.3
urllib3==2.2.2
uvicorn==0.11.3
uvloop==0.19.0
virtualenv==20.26.3
wadllib==1.3.3
webencodings==0.5.1
websockets==8.1
Werkzeug==3.0.4
wsaccel==0.6.2
wsproto==0.15.0
xhtml2pdf==0.2.16
zeep==4.2.1
zipp==3.20.2
zope.component==4.3.0
zope.event==4.4
zope.hookable==5.0.0
zope.interface==4.7.1
